Refactor: Reorganize and enhance CLI argument definitions, improve grouping and clarity in `CommandLineParser`, and use `size_t` for type safety in `TypeManager`.

+++ src/Compiler/Sema/Type/TypeManager.cpp
@@ -376,7 +376,7 @@ void TypeManager::buildPromoteTable()
         typeChar_,
         typeRune_};
 
-    constexpr uint32_t n = types.size();
+    constexpr size_t n = types.size();
 
     promoteIndex_.clear();
     promoteIndex_.reserve(n);
